fix a build error in the performance test

+++ perf-tests/nanoflann/test_nanoflann.cpp
@@ -145,7 +145,7 @@ void perf_test(const size_t N)
 	t0=get_time();
 
 	const size_t num_results = 1;
-	int ret_index;
+	size_t ret_index;
 	num_t out_dist_sqr;
 	nanoflann::KNNResultSet<num_t> resultSet(num_results);
 	resultSet.init(&ret_index, &out_dist_sqr );
